Replace markup parser with new one
Drop in the new parser.
